Move most xsbti implementations under dedicated namespace
