+++ src/main/scala/coursier/sbtlauncher/implem/ScalaProvider.scala
@@ -1,4 +1,4 @@
-package coursier.sbtlauncher
+package coursier.sbtlauncher.implem
 
 import java.io.File
